Only overwrite configure script if necessary
The script which configures a file at build time would be newer than the
output file which was only opportunistically overwritten which caused
make to be unhappy after a reconfigure of sprokit because the script was
newer than the output file but it was never actually updated. To fix
this, opportunistically write the script as well.